requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0

# Experiment tracking
wandb>=0.15.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Data manipulation
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing
pillow>=9.5.0

# For CLIP/BLIP experiments (Phase 2-3)
transformers>=4.30.0

# Utilities
scikit-learn>=1.3.0

# For table formatting in results
tabulate>=0.9.0
